+++ test/3estPrimitiveServer/PrimitiveServer.cpp
@@ -313,7 +313,7 @@ std::ostream &operator<<(std::ostream &o, const Matrix4f &transform)
 {
   o << "[\n";
 
-  for (int i = 0; i < 4; ++i)
+  for (size_t i = 0; i < 4; ++i)
   {
     if (i > 0)
     {
